COIN-928: Added Redshift provisioned/serverless support
Added support for Redshift Provisioned and Serverless. The package will
check which adapter is running to figure out if it's Redshift and then
combine it with a variable to determien if it's Serverless or not.
Ideally the adapter would also tell us that but it doesn't currently.

For Provisioned clusters, the user will provide the cost per node in the
cluster as well as the number of nodes to estimate the cost. Queries are
then priced based on how much time they used of the cluster. This isn't
perfect because queries can run concurrently so we somewhat over
estimate the costs.

For Serverless clusters, the user will provde the cost per RPU hour
consumed. We then estimate how many RPU's a query consumed and then
combine those two pieces together.